+++ CHANGES.md
@@ -1,3 +1,9 @@
+Unreleased
+--------------
+
+- archi: add `Component.identity`which creates a component whose value is its
+  argument ([#11](https://github.com/anmonteiro/archi/pull/11))
+
 0.1.0 2020-05-07
 --------------